Empty existing implementation of ExhaustiveQueryGraphSolver
